Moving classes/* out to src as these aren't really classes anymore.
