+++ ui/guest/src/elementRegistry.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2020 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -14,8 +14,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import * as iceRegistry from './ICERegistry';
-import { findContainerRecord, getById } from './ICERegistry';
+import * as iceRegistry from './iceRegistry';
+import { findContainerRecord, getById } from './iceRegistry';
 import {
   byPathFetchIfNotLoaded,
   getCachedContentType,
@@ -23,7 +23,7 @@ import {
   hasCachedModel,
   isInheritedField,
   model$
-} from './ContentController';
+} from './contentController';
 import { take } from 'rxjs/operators';
 import * as ContentType from '@craftercms/studio-ui/utils/contentType';
 import * as Model from '@craftercms/studio-ui/utils/model';
@@ -33,13 +33,13 @@ import {
   ElementRecordRegistration,
   HighlightData,
   ICERecord
-} from '../models/InContextEditing';
+} from './models/InContextEditing';
 import { ValidationResult } from '@craftercms/studio-ui/models/ContentType';
-import { RegistryEntry } from '../models/Registry';
+import { RegistryEntry } from './models/Registry';
 import { LookupTable } from '@craftercms/studio-ui/models/LookupTable';
 import { notNullOrUndefined, nullOrUndefined } from '@craftercms/studio-ui/utils/object';
 import { forEach } from '@craftercms/studio-ui/utils/array';
-import { getChildArrangement, sibling } from '../utils/dom';
+import { getChildArrangement, sibling } from './utils/dom';
 import $ from 'jquery';
 
 let seq = 0;
